% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/distrcox.R
\name{coxAnonymize}
\alias{coxAnonymize}
\title{Function for anonymizing a \code{\link[survival]{coxph}} object}
\usage{
coxAnonymize(model)
}
\arguments{
\item{model}{A \code{\link[survival]{coxph}} object}
}
\value{
An anonymized \code{\link[survival]{coxph}} object
}
\description{
Function that anonymizes a \code{\link[survival]{coxph}} class object by deleting each reference to single patients data.
}
\details{
The data to be anonymized are in the structure of the \code{\link[survival]{coxph}} object. They are:
\itemize{
   \item linear.predictors
   \item residuals
   \item nevent
   \item y
} An anonymized object can be transmitted far from the origin without giving the chance to reconstruct the single cases details.
}

