% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/distrcox.R
\name{coxSimulate}
\alias{coxSimulate}
\title{Function for simulating a Cox distribution series}
\usage{
coxSimulate(n = 1000, input.model, features, BH)
}
\arguments{
\item{n}{Number of cases to be simulated}

\item{input.model}{An object of \code{\link[survival]{coxph}} class, even anonymized (see \code{\link{coxAnonymize}})}

\item{features}{An object of class \code{features} containing description of single covariates in the dataset}

\item{BH}{Cumulative baseline hazard as calculated by \code{\link[survival]{basehaz}} function applied to a \code{\link[survival]{coxph}} object}
}
\value{
A \code{\link[base]{data.frame}} containing simulated observation
}
\description{
Function for simulating a Cox distribution series
}
\details{
Cox distribution series are based on the assumption of \emph{proportional hazards} of covariate
over a given underneath \emph{baseline hazards} function. The hypotesis of proportionality is essential
for achieving good performance by models generated by this kind of fitting. The \code{coxSimulate} function
works on an input \code{\link[survival]{coxph}} model. The model can be also anonymized by \code{\link{coxAnonymize}}
function, in order to use the simulation function in a \emph{distributed learning} context, without moving data from
their site of origin and without carrying details of single patients within the \code{\link[survival]{coxph}} object that
might even hypothetically be used to reproduce the original dataset coming grom its own (far) source. \emph{Baseline hazard} has
to be given in order to produce a cox simulation because it cannot be anymore calculated by using the \code{\link[survival]{coxph}}
}

