% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/distrcox.R
\name{extractFeatures}
\alias{extractFeatures}
\title{Function for extracting summary of covariates in a \code{\link[base]{data.frame}} object}
\usage{
extractFeatures(dataset, breaks = 20)
}
\arguments{
\item{dataset}{A \code{data.frame} object}

\item{breaks}{Numnber of \code{breaks} in the \code{hist} objects that returns the summary of \code{numeric} covariates}

\item{times}{Numeric vector of observation times or death in the dataset}

\item{status}{Numeric vector of censored \code{(0)} or event \code{(1)} cases in the dataset}
}
\value{
An object of class \code{features} with following structure:
\itemize{
   \item \code{classes}. A character vector with names of classes of each covariate.
   \item \code{prop}. A list of \code{table} objects for integer and factor covaraites, or \code{hist} objects for numeric covariates.
   \item \code{numCases}. The total number of cases in the dataset.
}
}
\description{
This function is used for building a summary list of the covariates in a \code{\link[base]{data.frame}} object. It has to be used before
using \code{\link{coxSimulate}} in order to generate new dataset with given baseline hazards and a set of coefficients set for a given outcome.
}
\details{
When a simulation of new Cox dataset is performed according the features of a given dataset a pre-analysis of the
 dataset has to be performed by using this function. The result given by \code{extractFeatures} can be directly put in \code{\link{coxSimulate}}
 function to achieve a new dataset similar to the one that gave the model used for \code{coeff}, \code{cov} and \code{BH} arguments.
}

