% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/distrcox.R
\name{extracth0}
\alias{extracth0}
\title{Function that extracts the hazard function from a \code{coxph} object}
\usage{
extracth0(dataset, model, times, events, smoothed = FALSE, spar = 0.75,
  grid = 1000)
}
\arguments{
\item{dataset}{The dataset as \code{data.frame} for modeling the \emph{baseline hazard rate}}

\item{model}{The \code{\link[survival]{coxph}} object for calculating the \emph{baseline hazard rate}}

\item{times}{The vector of times where \emph{events} are recorded}

\item{events}{The vector of the events \code{(0)} for censored cases, \code{(1)} for recorded events}

\item{smoothed}{If \code{TRUE} return values smoothed according a \code{\link[stats]{smooth.spline}} function}

\item{spar}{Smoothing parameter, used only if \code{smoothed = TRUE}}

\item{grid}{Number of estimation point of \eqn{h_0(t)} function}
}
\value{
A \code{data.frame} with \eqn{h_0} values and related times.
}
\description{
This function extracts the hazard function \eqn{h_0} from a given \code{\link[survival]{coxph}} object. The object must be provided with
the original dataset used for the modeling. The result of the function isn't equivalent to the result of \code{\link[survival]{basehaz}} function, because
the latter provides the \emph{cumulative baseline hazard} \eqn{H_0(t)} and not the \emph{hazard function} itself.
}
\details{
The survival function of a Cox proportional hazards model is given by the equation:
\deqn{S\left ( t|x \right )=\exp \left [ -H_0\left ( t \right )*\exp\left ( \beta'X \right ) \right ]}
where:
\deqn{H_0\left ( t \right )=\int_{0}^{t}h_0\left ( u \right )du}
Knowing the value of \eqn{h_0\left ( t \right )} it is possible for a given value of \eqn{\beta'X} to compute the expected survival time. In fact given:
\deqn{h(t)=h_0(t)*\exp\left ( \beta'X \right )}
and censoring time \eqn{c=U*max(t)} where \eqn{U \sim Uni\left [ 0,1 \right ]}, that is \eqn{U} follows a uniform distribution on the interval from 0 to 1,
and \eqn{max(t)} is the maximum observation time for the population to be simulated, the calculated survival time is \deqn{T=\frac{-\log{\left (U  \right )}}{h_t}}
It is possible to simulate the \emph{censoring} in the population by considering that censored cases are they where \eqn{c<T}.
Starting from a base \code{\link[survival]{coxph}} model it is possible to give in output a matrix of values of time and \eqn{h_0(t)} needed for the simulation
of a Cox distribution with the same covariates, coefficients and baseline hazard of the generating one.
}

