% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/distrcox.R
\name{HR.fit}
\alias{HR.fit}
\title{Function for fitting the baseline hazard function}
\usage{
HR.fit(coxobj, init = c(0.1, 0.1))
}
\description{
The function fits a baseline hazard model according a value of discrete baseline hazards as function of time \eqn{t} achieved by using
\code{\link[survival]{basehaz}} function over a \code{\link[survival]{coxph}} object.
}

