% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/distrcox.R
\name{HR.gompertz}
\alias{HR.gompertz}
\title{Function for calculating Gompertz based hazard rate H0}
\usage{
HR.gompertz(t, lambda, alpha, inv = FALSE)
}
\arguments{
\item{t}{time}

\item{lambda}{\eqn{\lambda} parameter of Gompertz function}

\item{alpha}{\eqn{\alpha} parameter of Gompertz equation}
}
\value{
A numeric vector of the baseline hazard values \eqn{H_{0}\left (t \right)} or their inverse \eqn{H_{0}^{-1}\left ( t \right )}.
}
\description{
Gompertz function describing the hazard rate as function of the time
}
\details{
Given a Cox fit object returned by a \code{\link[survival]{coxph}} function this function calulates the parameters to achieve the value of baseline hazard rate
as function of the time. The result returned by function is either \deqn{H_0 \left ( t \right )=\left [\frac{1}{\alpha}log\left ( \frac{\alpha}{\lambda}t+1 \right )  \right ]^{-1}}
or \deqn{\frac{1}{H_0 \left ( t \right )}=\frac{1}{\alpha}log\left ( \frac{\alpha}{\lambda}t+1 \right )}
}

