% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/distrcox.R
\docType{data}
\name{rectum}
\alias{rectum}
\title{Beijing post-operative rectal cancer patients dataset}
\format{A data frame with 1798 rows and 17 variables}
\usage{
rectum
}
\description{
A dataset containing the Beijing University data about post-operative radiotherapy in rectal cancer patients. The recorded features are:
\itemize{
  \item Age. Age of patient
  \item Sex. Male, Female
  \item ECOG. Pre-treatment performance status
  \item Histology. Adenocarcinoma, Mucinous
  \item Grade. Histological differentiation grade
  \item RTCT. If treated by chemoradiotherapy (1) or radiotherapy alone (0)
  \item pT. Pathological T stage
  \item pN. Pathological N stage
  \item pM. Pathological M stage
  \item AdjCT. If patient underwent (1) or not (0) to adjuvant chemotherapy
  \item CTcicles. Number of cicles of adjuvant chemotherapy
  \item RTbreak. Duration (in days) of breaks during radiotherapy treatment
  \item RTtox. Maximum grade of RTOG toxicity shown during (chemo)radiotherapy treatment
  \item Residual. Presence (1) or absence (0) of residual tumor after surgery
  \item IMRT. If radiotherapy treatment was delivered by IMRT (1) or not (0)
  \item status. Censored cases (0) or dead (1)
  \item OS. Overall survival time (in months)
}
}
\author{
Liu Wen Yang \email{liuwenyang26@163.com}

Nicola Dinapoli \email{nicola.dinapoli@rm.unicatt.it}
}
\keyword{data}

