% Generated by roxygen2 (4.0.2): do not edit by hand
\name{coxSimulate}
\alias{coxSimulate}
\title{Function for simulating a Cox distirbution series}
\usage{
coxSimulate(n = 1000, coeff)
}
\arguments{
\item{n}{Number of cases to be simulated}

\item{coeff}{Coefficients achieved by a \code{\link[survival]{coxph}} object}
}
\value{
A list containing a matrix with cases in lines ordered by coefficients as got by coeff param and corresponding coxph model
}
\description{
Function for simulating a Cox distribution series
}
\details{
Cox distribution series are based on the assumption of \emph{proportional hazards} of covariate
over a given underneath \emph{baseline hazards} function. The hypotesis of proportionality is essential
for achieving good performance by models achieved by this kind of fitting. When the verification of
proportionality is needed the function \code{\link[survival]{cox.zph}} should be used to perform the
verification.
}

