% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/distrcox.R
\name{HR.weibull}
\alias{HR.weibull}
\title{Function for calculating Weibull based hazard rate H0}
\usage{
HR.weibull(t, lambda, ni, inv = FALSE)
}
\arguments{
\item{t}{time}

\item{lambda}{\eqn{\lambda} parameter of Weibull function}

\item{ni}{\eqn{\nu} parameter of Weibull equation}

\item{inv}{if \code{FALSE} returns the inverse value of baseline hazard, if \code{TRUE} returns the baseline hazard.}
}
\value{
A numeric vector of the baseline hazard values \eqn{H_{0}\left (t \right)} or their inverse \eqn{H_{0}^{-1}\left ( t \right )}.
}
\description{
Weibull function describing the hazard rate as function of the time.
}
\details{
Given a Cox fit object returned by a \code{\link[survival]{coxph}} function this function calulates the parameters to achieve the value of baseline hazard rate
as function of the time. The result returned by function is either \deqn{\frac{1}{H_0 \left ( t \right )}=\left ( \lambda^{-1}t \right )^{\frac{1}{\nu}}} or
\deqn{H_0 \left ( t \right )=\left ( \lambda^{-1}t \right )^{-\frac{1}{\nu}}}
}

